Fix PT_PARENT property type to compile with recent wxWidgets versions
A recent change in wxWidgets made wxPGProperty an abstract class so that it can't be used directly anymore. Use a read-only wxStringProperty instead to get almost the previous behavior. This is not ideal but the previous version already did not honor the property grid "style" of composed properties.

The usage of the properties should be examined and maybe their implementation adjusted to better use the features of the current property grid version.